Moved the DataTable component to the containers directory. #11
