+++ client/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { Container } from 'reactstrap'
 import Header from './components/Header'
 import Wizard from './containers/Wizard'
-import DataTable from './components/DataTable'
+import DataTable from './containers/DataTable'
 import forecastAnnualEquity from './util/calculations'
 import Chart from './containers/Chart'
 import Summary from './containers/Summary'
